+++ rust/src/libponyc/pkg/use.rs
@@ -1121,7 +1121,7 @@ pub unsafe extern "C" fn use_command(
         ast,
         (::core::mem::size_of::<[*mut *mut ast_t; 4]>() as libc::c_ulong)
             .wrapping_div(::core::mem::size_of::<*mut *mut ast_t>() as libc::c_ulong)
-            .wrapping_sub(1 as libc::c_int as libc::c_ulong),
+            .wrapping_sub(1),
         children.as_mut_ptr(),
     );
     if !spec.is_null() {
